I quit the game jam
I have decided to back out from the jam. This game turned into a passion project for me and I want to see it through without rushing it. I believe it's already the most beautiful video game I've ever made but I know it can truly blossom if I spend more time on it. My current scope is to finish and publish the game by the end of 2024. That means I have just a little over four (4) months to make it. It's going to be a beautiful story and an engaging gaming experience.